requirements.txt
streamlit
langchain-community
langchain-groq
pypdf
python-dotenv
ipykernel
PyPDF2